+++ gradle.properties
@@ -42,7 +42,7 @@ spring-jcl=5.3.15
 log4j2.version=2.17.1
 assetPluginVersion=3.4.3
 dbMigrationPluginVersion=4.0.0-RC2
-node.install=12.16.3
+node.install=16.13.0
 org.gradle.daemon=true
 org.gradle.parallel=true
 org.gradle.jvmargs=-Dfile.encoding=UTF-8 -Xmx1024M
