+++ rundeckapp/grails-spa/packages/ui-trellis/babel.config.js
@@ -1,7 +1,5 @@
 module.exports = {
-    presets: [
-        ["@vue/cli-plugin-babel/preset", {
-            targets: {node: "current"}
-        }]
-    ]
-};
+  presets: [
+    "@vue/cli-plugin-babel/preset"
+  ]
+}
