+++ rundeckapp/grails-spa/packages/ui-trellis/build.gradle
@@ -29,7 +29,7 @@ node{
     
     // Version of node to download and install (only used if download is true)
     // It will be unpacked in the workDir
-    version = "12.16.3"
+    version = "v16.13.0"
     
     // // Version of npm to use
     // // If specified, installs it in the npmWorkDir
