+++ rundeckapp/grails-spa/packages/ui-trellis/jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
-    preset: '@vue/cli-plugin-unit-jest/presets/typescript-and-babel',
-    testEnvironment: 'node',
-    modulePathIgnorePatterns: ['public']
-  }
-  
+  preset: '@vue/cli-plugin-unit-jest/presets/typescript-and-babel',
+  testEnvironment: 'node',
+  modulePathIgnorePatterns: ['public']
+}
