+++ rundeckapp/grails-spa/packages/ui-trellis/package.json
@@ -11,8 +11,7 @@
     "build": "run-script-os",
     "build:darwin:linux": "./node_modules/.bin/vue-cli-service build --mode production && tsc -p ./tsconfig.build.json",
     "build:win32": "vue-cli-service.cmd build --mode production && tsc -p ./tsconfig.build.json",
-    "storybook": "start-storybook -s public,theme,tests/data",
-    "storybook:next": "STORYBOOK_THEME=theme-next start-storybook -s public,theme-next,tests/data"
+    "storybook": "start-storybook -s public,theme,tests/data"
   },
   "repository": {
     "type": "git",
@@ -25,86 +24,74 @@
   },
   "homepage": "https://github.com/rundeck/ui-trellis#readme",
   "dependencies": {
-    "@popperjs/core": "^2.8.4",
+    "@rundeck/client": "^0.2.3",
+    "ace-builds": "^1.4.12",
+    "axios": "^0.18.0",
+    "core-js": "^3.6.5",
+    "markdown-it-vue": "^1.1.5",
+    "mobx": "^5.15.4",
+    "moment": "^2.25.3",
+    "perfect-scrollbar": "^1.5.0",
+    "prex": "^0.4.7",
+    "uiv": "^0.33.0",
+    "vue": "^2.6.11",
+    "vue-good-table": "^2.21.11",
+    "vue-i18n": "^8.22.4",
+    "vue-multiselect": "^2.1.6",
+    "vue-property-decorator": "^9.1.2",
+    "vue-router": "^3.3.2",
+    "vue-tsx-support": "^2.3.0",
+    "vue-virtual-scroller": "^1.0.10",
+    "webpack-node-externals": "^3.0.0"
+  },
+  "devDependencies": {
     "@storybook/addon-backgrounds": "^6.2.9",
     "@storybook/addon-docs": "^6.2.9",
     "@storybook/addon-knobs": "^6.2.9",
     "@storybook/preset-typescript": "^3.0.0",
     "@storybook/vue": "^6.2.9",
-    "babel-preset-vue": "^2.0.2",
-    "deep-object-diff": "1.1.0",
-    "perfect-scrollbar": "^1.5.0",
-    "storybook-dark-mode": "^1.0.8",
-    "uuid": "8.3.2",
-    "vue-good-table": "^2.21.10",
-    "vue-i18n": "^8.22.4",
-    "vue-multiselect": "^2.1.6",
-    "vue-virtual-scroller": "^1.0.10"
-  },
-  "devDependencies": {
-    "@azure/ms-rest-js": "^2.0.7",
-    "@babel/core": "^7.10.2",
-    "@rundeck/client": "0.2.3",
-    "@types/node": "14.0.13",
+    "@types/jest": "^24.0.19",
+    "@types/node": "^14.0.13",
     "@types/uuid": "^8.3.0",
-    "@vue/cli-plugin-babel": "4.5.8",
-    "@vue/cli-plugin-typescript": "4.5.8",
-    "@vue/cli-plugin-unit-jest": "4.5.8",
-    "@vue/cli-service": "4.5.8",
-    "ace-builds": "1.4.12",
-    "axios": "^0.18.0",
-    "axios-mock-adapter": "^1.18.1",
-    "babel-loader": "^8.1.0",
-    "concurrently": "^5.2.0",
-    "deep-object-diff": "1.1.0",
+    "@vue/cli-plugin-babel": "^4.5.8",
+    "@vue/cli-plugin-typescript": "^4.5.8",
+    "@vue/cli-plugin-unit-jest": "^4.5.8",
+    "@vue/cli-service": "^4.5.8",
+    "@vue/test-utils": "^1.0.3",
+    "concurrently": "^7.0.0",
     "fetch-mock": "^9.10.1",
-    "fs-extra": "9.0.1",
-    "isomorphic-fetch": "^2.2.1",
-    "less-loader": "^6.1.0",
-    "markdown-it-vue": "1.1.5",
-    "mobx": "^5.15.4",
     "mobx-utils": "^5.6.1",
     "mobx-vue": "^2.0.11",
-    "moment": "^2.25.3",
-    "node-fetch": "^2.6.0",
-    "node-sass": "^4.14.1",
-    "prex": "^0.4.7",
-    "run-script-os": "^1.1.1",
+    "run-script-os": "^1.1.6",
+    "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
-    "typescript": "4.2.3",
-    "uiv": "0.33.0",
-    "vue-class-component": "7.2.3",
-    "vue-moment": "^4.1.0",
-    "vue-property-decorator": "8.5.1",
-    "vue-router": "^3.3.2",
-    "walk": "^2.3.14",
-    "webpack-node-externals": "^1.7.2"
+    "storybook-dark-mode": "^1.0.8",
+    "typescript": "^4.2.3",
+    "uuid": "^8.3.2",
+    "vue-class-component": "^7.2.3",
+    "vue-cli-plugin-tsx-support": "~0.4.1",
+    "vue-jsx-hot-loader": "^1.4.1",
+    "vue-template-compiler": "^2.6.11",
+    "vue-tsx-support": "^3.2.0",
+    "walk": "^2.3.15"
   },
   "peerDependencies": {
-    "@rundeck/client": "0.2.3",
-    "ace-builds": "1.4.12",
+    "@rundeck/client": "^0.2.3",
+    "ace-builds": "^1.4.12",
     "axios": "^0.18.0",
-    "markdown-it-vue": "^1.0.11",
+    "markdown-it-vue": "^1.1.5",
     "mobx": "^5.15.4",
     "mobx-utils": "^5.6.1",
     "mobx-vue": "^2.0.11",
     "moment": "^2.25.3",
     "prex": "^0.4.7",
-    "uiv": "0.33.0",
-    "vue": "2.6.12",
+    "uiv": "^0.33.0",
+    "vue": "^2.6.11",
     "vue-class-component": "^7.2.3",
-    "vue-moment": "^4.1.0",
-    "vue-property-decorator": "^8.4.1",
+    "vue-property-decorator": "^9.1.2",
     "vue-router": "^3.3.2"
   },
   "publishConfig": {
     "access": "public"
-  },
-  "files": [
-    ".storybook",
-    "lib",
-    "src",
-    "theme",
-    "theme-next"
-  ]
+  }
 }
