+++ rundeckapp/grails-spa/packages/ui-trellis/src/ markdown-it-vue.d.ts
@@ -1,0 +1,2 @@
+// markdown-it-vue.d.ts
+declare module 'markdown-it-vue';
