+++ rundeckapp/grails-spa/packages/ui-trellis/src/App.vue
@@ -1,0 +1,33 @@
+<template>
+  <div id="app">
+
+ 
+    <copybox/>
+  </div>
+</template>
+
+<script lang="ts">
+import Vue from 'vue';
+import HelloWorld from './components/HelloWorld.vue';
+import copybox from './components/containers/copybox/CopyBox.vue';
+
+
+export default Vue.extend({
+  name: 'App',
+  components: {
+    HelloWorld,
+    copybox,
+  }
+});
+</script>
+
+<style lang="scss">
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
+}
+</style>
