--- rundeckapp/grails-spa/packages/ui-trellis/src/antScope.less
@@ -1,1 +1,0 @@
-#antd { @import '~ant-design-vue/lib/style/core/index.less'; @import '~ant-design-vue/lib/style/themes/default.less'; }
