+++ rundeckapp/grails-spa/packages/ui-trellis/src/shims-tsx.d.ts
@@ -1,0 +1,15 @@
+import 'vue-tsx-support/enable-check';
+
+import Vue, { VNode } from 'vue'
+
+declare global {
+  namespace JSX {
+    // tslint:disable no-empty-interface
+    interface Element extends VNode {}
+    // tslint:disable no-empty-interface
+    interface ElementClass extends Vue {}
+    interface IntrinsicElements {
+      [elem: string]: any;
+    }
+  }
+}
