+++ rundeckapp/grails-spa/packages/ui-trellis/src/shims-vue.d.ts
@@ -1,0 +1,11 @@
+declare module '*.vue' {
+  import Vue from 'vue';
+  export default Vue;
+}
+
+declare module 'vue2-ace-editor' {
+
+}
+declare module 'markdown-it-vue'{
+
+}
