--- rundeckapp/grails-spa/packages/ui-trellis/src/vue-shims.d.ts
@@ -1,11 +1,0 @@
-declare module '*.vue' {
-  import Vue from 'vue';
-  export default Vue;
-}
-
-declare module 'vue2-ace-editor' {
-
-}
-declare module 'markdown-it-vue'{
-
-}
