+++ rundeckapp/grails-spa/packages/ui-trellis/tsconfig.json
@@ -1,28 +1,28 @@
 {
-    "compilerOptions": {
-        "outDir": "lib",
-        "baseUrl": ".",
-        "sourceMap": true,
-        "strict": true,
-        "moduleResolution": "node",
-        "target": "esnext",
-        "experimentalDecorators": true,
-        "skipLibCheck": true,
-        "esModuleInterop": true,
-        "declaration": true,
-        "jsx": "preserve",
-        "allowSyntheticDefaultImports": true,
-        "types": [
-            "webpack-env",
-            "jest"
-        ],
-        "paths": {
-            "react": [".sink.d.ts"]
-        }
-    },
-    "include": [
-        "./src/**/*",
-        "tests/**/*.ts",
-        "tests/**/*.tsx"
-    ]
+  "compilerOptions": {
+      "outDir": "lib",
+      "baseUrl": ".",
+      "sourceMap": true,
+      "strict": true,
+      "moduleResolution": "node",
+      "target": "esnext",
+      "experimentalDecorators": true,
+      "skipLibCheck": true,
+      "esModuleInterop": true,
+      "declaration": true,
+      "jsx": "preserve",
+      "allowSyntheticDefaultImports": true,
+      "types": [
+          "webpack-env",
+          "jest"
+      ],
+      "paths": {
+          "react": [".sink.d.ts"]
+      }
+   },
+   "include": [
+         "./src/**/*",
+         "tests/**/*.ts",
+         "tests/**/*.tsx"
+   ]
 }
