+++ rundeckapp/grails-spa/packages/ui/package.json
@@ -34,6 +34,7 @@
     "mobx-vue": "^2.0.11",
     "moment": "^2.25.3",
     "prex": "0.4.7",
+    "sass": "^1.49.7",
     "source-map-loader": "^1.0.1",
     "uiv": "^0.33.0",
     "vue": "2.6.14",
@@ -66,7 +67,6 @@
     "lerna": "^3.4.3",
     "less-loader": "^6.1.0",
     "node-notifier": "^5.1.2",
-    "node-sass": "^4.12.0",
     "ora": "^1.2.0",
     "pnp-webpack-plugin": "^1.6.4",
     "portfinder": "^1.0.13",
